+++ test/TSmartAccount7702/ValidateUserOp.t.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.33;
 import {MockEntryPoint} from "../mocks/MockEntryPoint.sol";
 import {UseEntryPointV09} from "./entrypoint/UseEntryPointV09.sol";
 import {SmartWalletTestBase} from "./SmartWalletTestBase.sol";
-import {SmartAccount7702} from "../../src/SmartAccount7702.sol";
+import {TSmartAccount7702} from "../../src/TSmartAccount7702.sol";
 import {PackedUserOperation} from "account-abstraction/interfaces/PackedUserOperation.sol";
 
 /// @dev Abstract test logic for validateUserOp(). Concrete classes provide the EntryPoint version.
@@ -47,7 +47,7 @@ abstract contract TestValidateUserOpBase is SmartWalletTestBase {
         userOp.signature = abi.encodePacked(bytes32(0), bytes32(0), uint8(27));
 
         // Calling directly (not from EntryPoint) should revert
-        vm.expectRevert(SmartAccount7702.Unauthorized.selector);
+        vm.expectRevert(TSmartAccount7702.Unauthorized.selector);
         account.validateUserOp(userOp, keccak256("123"), 0);
     }
 
@@ -123,7 +123,7 @@ abstract contract TestValidateUserOpBase is SmartWalletTestBase {
 
         uint256 recipientBefore = recipient.balance;
 
-        userOpCalldata = abi.encodeCall(SmartAccount7702.execute, (recipient, 1 ether, ""));
+        userOpCalldata = abi.encodeCall(TSmartAccount7702.execute, (recipient, 1 ether, ""));
 
         // Build UserOp with non-zero gas fees and no paymaster
         PackedUserOperation memory userOp = PackedUserOperation({