+++ test/gas/EndToEnd.t.sol
@@ -4,12 +4,12 @@ pragma solidity ^0.8.33;
 import {console2} from "forge-std/Test.sol";
 import {PackedUserOperation} from "account-abstraction/interfaces/PackedUserOperation.sol";
 
-import {SmartAccount7702} from "../../src/SmartAccount7702.sol";
+import {TSmartAccount7702} from "../../src/TSmartAccount7702.sol";
 import {MockERC20} from "../mocks/MockERC20.sol";
 
 import {MockTarget} from "../mocks/MockTarget.sol";
-import {SmartWalletTestBase} from "../SmartAccount7702/SmartWalletTestBase.sol";
-import {UseEntryPointV09} from "../SmartAccount7702/entrypoint/UseEntryPointV09.sol";
+import {SmartWalletTestBase} from "../TSmartAccount7702/SmartWalletTestBase.sol";
+import {UseEntryPointV09} from "../TSmartAccount7702/entrypoint/UseEntryPointV09.sol";
 
 /// @title EndToEndTest
 /// @notice Gas comparison tests between ERC-4337 Base Account and EOA transactions
@@ -43,7 +43,7 @@ contract EndToEndTest is SmartWalletTestBase, UseEntryPointV09 {
         uint256 recipientBefore = address(0x1234).balance;
 
         // Prepare UserOperation for native ETH transfer
-        userOpCalldata = abi.encodeCall(SmartAccount7702.execute, (address(0x1234), 1 ether, ""));
+        userOpCalldata = abi.encodeCall(TSmartAccount7702.execute, (address(0x1234), 1 ether, ""));
         PackedUserOperation memory op = _getUserOpWithSignature();
 
         // Measure calldata size
@@ -87,7 +87,7 @@ contract EndToEndTest is SmartWalletTestBase, UseEntryPointV09 {
 
         // Prepare UserOperation for ERC20 transfer
         userOpCalldata = abi.encodeCall(
-            SmartAccount7702.execute, (address(usdc), 0, abi.encodeCall(usdc.transfer, (address(0x5678), 100e6)))
+            TSmartAccount7702.execute, (address(usdc), 0, abi.encodeCall(usdc.transfer, (address(0x5678), 100e6)))
         );
         PackedUserOperation memory op = _getUserOpWithSignature();